refactor: chess 패키지 추가 후 디렉토리 이동
