+++ src/main/java/softeer2nd/chess/utils/StringUtils.java
@@ -1,4 +1,4 @@
-package softeer2nd.utils;
+package softeer2nd.chess.utils;
 
 public class StringUtils {
     public static final String NEWLINE = System.getProperty("line.separator");
