+++ src/test/java/softeer2nd/chess/domain/BoardTest.java
@@ -1,9 +1,11 @@
-package softeer2nd.domain;
+package softeer2nd.chess.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import softeer2nd.domain.enums.Color;
+import softeer2nd.chess.domain.Board;
+import softeer2nd.chess.domain.Piece;
+import softeer2nd.chess.domain.enums.Color;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
